CURB-4299 Extracted logic of ResponsiveContainer component to reusable hook
